// client/src/interfaces/ClientErrors.interface.ts
export default interface IClientErrors {
  name?: string;
  email?: string;
  password?: string;
  password2?: string;
  badRequest?: string;
  notWorking?: string;
  noProfile?: string;
  noUser?: string;
  noPost?: string;
  noPosts?: string;
  noProfiles?: string;
  notAuthorized?: string;
  notLiked?: string;
  commentNotExits?: string;
  alreadyLiked?: string;
  userId?: string;
  twitter?: string;
  facebook?: string;
  linkedin?: string;
  youtube?: string;
  instagram?: string;
  handle?: string;
  status?: string;
  company?: string;
  website?: string;
  location?: string;
  skills?: string;
  githubusername?: string;
  bio?: string;
  description?: string;
  to?: string;
  from?: string;
  title?: string;
  school?: string;
  degree?: string;
  fieldofstudy?: string;
  text?: string;
}